# contracts/openapi/topsdojob-v3-local.yaml
openapi: 3.1.0
info:
  title: Tops do Job V3 Local API
  version: 0.0.1
  description: "Contrato minimo local do skeleton, camada transversal, API publica de leitura e funil local Anuncie gratis. A API segue fail-closed: endpoints nao descritos em /api/** nao devem ficar publicos por acidente."
servers:
  - url: http://localhost:8080
    description: backend local
paths:
...
  /api/admin/anuncios/{id}:
    get:
      summary: Detalhe administrativo sanitizado de anuncio
      description: Retorna os dados comerciais necessarios a moderacao, proprietario mascarado e revisao aberta, somente para ADMIN ou MODERADOR com ANUNCIO_LER.
      operationId: getAdminAnuncioDetalhe
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: detalhe sanitizado de anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnuncioDetalhe"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/desempenho/anuncios/{id}:
    get:
      summary: Desempenho read-only de anuncio
      description: Retorna prova de resultado local e sanitizada para um anuncio, com visualizacoes, cliques WhatsApp permitidos, taxa clique/view, origens agregadas e comparativo organico/Premium. Nao expoe IP, User-Agent, hash interno, referer bruto, telefone/WhatsApp bruto, storage, pagamento ou credito. Nao executa pixel, tracking externo, cobranca, impulsionamento ou mutacao.
      operationId: getAdminDesempenhoAnuncio
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: desempenho sanitizado do anuncio
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDesempenhoAnuncio"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/desempenho/anuncios/{id}/diario:
    get:
      summary: Serie diaria de desempenho read-only
      description: Lista agregados diarios locais de visualizacoes e cliques WhatsApp, marcando dias com beneficio Premium ativo. Nao retorna eventos brutos, IP, User-Agent, hash interno ou referer bruto.
      operationId: getAdminDesempenhoAnuncioDiario
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: serie diaria sanitizada
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDesempenhoDiario"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/desempenho/anuncios/{id}/origens:
    get:
      summary: Origens agregadas de desempenho read-only
      description: Retorna origem, cidade, bairro, UF e totais agregados de visualizacoes/cliques. Os dados sao sanitizados e nao expoem referer bruto, IP, User-Agent, hash ou identificador de visitante.
      operationId: getAdminDesempenhoAnuncioOrigens
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: origens agregadas sanitizadas
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDesempenhoOrigem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/desempenho/anunciantes/{usuarioId}:
    get:
      summary: Desempenho agregado por anunciante read-only
      description: Retorna desempenho agregado local e sanitizado por anunciante para uso administrativo futuro. Mantem endpoint real de anunciante indisponivel nesta fase e nao expoe contato, documento, pagamento, credito ou dado pessoal bruto.
      operationId: getAdminDesempenhoAnunciante
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: usuarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: desempenho agregado do anunciante
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDesempenhoAnunciante"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/desempenho/resumo:
    get:
      summary: Resumo agregado de desempenho read-only
      description: Retorna totais locais agregados de anuncios com metricas, visualizacoes, cliques WhatsApp permitidos e comparativo Premium. Nao garante resultado, nao limita gratuito, nao executa tracking externo e nao dispara acao comercial.
      operationId: getAdminDesempenhoResumo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: resumo agregado sanitizado
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDesempenhoResumo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    PublicSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao publica emitida pelo login; cookie HttpOnly, Secure em HML e SameSite Lax.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao administrativa local emitida por `POST /api/admin/auth/login`; cliente deve usar credentials include e o corpo da resposta nunca retorna token ou cookie.
  headers:
    RequestId:
      description: Identificador opaco de correlação da requisição.
      schema:
        type: string
        example: 11111111-1111-4111-8111-111111111111
  parameters:
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Identificador opaco enviado pelo cliente para correlação.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]{8,128}$"
...
  responses:
...
    NotFound:
      description: Recurso publico nao encontrado.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Autenticacao necessaria ou credenciais invalidas.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Usuario autenticado sem permissao suficiente.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    InternalError:
      description: Falha inesperada sem detalhes internos.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
...
    AdminLocalizacaoSanitizada:
      type: object
      additionalProperties: false
      properties:
        uf:
          type:
            - string
            - "null"
        cidade:
          type:
            - string
            - "null"
        bairro:
          type:
            - string
            - "null"
    AdminAnuncianteResumo:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        nome:
          type:
            - string
            - "null"
        emailMascarado:
          type:
            - string
            - "null"
          description: Identificador visual mascarado; nunca retorna o e-mail integral.
        status:
          type:
            - string
            - "null"
    AdminRevisaoAberta:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
        status:
          type: string
          enum: [ABERTA, EM_ANALISE]
        criadoEm:
          type: string
          format: date-time
    AdminAnuncioListaItem:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        slug:
          type:
            - string
            - "null"
        titulo:
          type:
            - string
            - "null"
        status:
          type:
            - string
            - "null"
        statusModeracao:
          type:
            - string
            - "null"
        localizacao:
          anyOf:
            - $ref: "#/components/schemas/AdminLocalizacaoSanitizada"
            - type: "null"
        criadoEm:
          type:
            - string
            - "null"
          format: date-time
        atualizadoEm:
          type:
            - string
            - "null"
          format: date-time
        publicadoEm:
          type:
            - string
            - "null"
          format: date-time
        midiasTotal:
          type:
            - integer
            - "null"
          format: int64
        revisoesTotal:
          type:
            - integer
            - "null"
          format: int64
        contatoConfigurado:
          type: boolean
          description: Indicador agregado; nao retorna telefone nem WhatsApp.
        documentoPendente:
          type: boolean
          description: Indicador agregado; nao retorna documento privado.
        comercialLimitado:
          type: boolean
        anunciante:
          anyOf:
            - $ref: "#/components/schemas/AdminAnuncianteResumo"
            - type: "null"
        revisaoAberta:
          anyOf:
            - $ref: "#/components/schemas/AdminRevisaoAberta"
            - type: "null"
    AdminAnuncioDetalhe:
      allOf:
        - $ref: "#/components/schemas/AdminAnuncioListaItem"
        - type: object
          additionalProperties: false
          properties:
            descricaoResumo:
              type:
                - string
                - "null"
              description: Texto sanitizado e mascarado; pode ser nulo quando ausente na origem.
            descricao:
              type:
                - string
                - "null"
              description: Conteudo comercial integral para moderacao; pode ser nulo quando ausente.
            categoria:
              type:
                - string
                - "null"
            ultimaPublicacaoEm:
              type:
                - string
                - "null"
              format: date-time
            precoInformado:
              type: boolean
            preco:
              type:
                - number
                - "null"
              format: double
              description: Preco informado no anuncio; pode ser nulo quando nao informado.
            whatsapp:
              type:
                - string
                - "null"
              description: Contato do proprio anuncio para analise; pode ser nulo quando nao informado.
            locaisAtendimento:
              type: array
              items:
                type: string
            servicos:
              type: array
              items:
                type: string
...
    AdminDesempenhoDiario:
      type: object
      additionalProperties: false
      properties:
        data:
          type: string
          format: date
        visualizacoes:
          type: integer
        cliquesWhatsapp:
          type: integer
        taxaCliqueView:
          type: number
          format: double
        premiumAtivo:
          type: boolean
        origemDados:
          type: string
          description: Fonte local agregada ou fallback local, sem evento bruto.
    AdminDesempenhoOrigem:
      type: object
      additionalProperties: false
      properties:
        origem:
          type: string
        uf:
          type:
            - string
            - "null"
        cidade:
          type:
            - string
            - "null"
        bairro:
          type:
            - string
            - "null"
        visualizacoes:
          type: integer
        cliquesWhatsapp:
          type: integer
        taxaCliqueView:
          type: number
          format: double
        dadosSensiveisOcultos:
          type: boolean
          const: true
    AdminDesempenhoComparativoPremium:
      type: object
      additionalProperties: false
      properties:
        visualizacoesComPremium:
          type: integer
        visualizacoesOrganicas:
          type: integer
        cliquesComPremium:
          type: integer
        cliquesOrganicos:
          type: integer
        taxaCliqueViewComPremium:
          type: number
          format: double
        taxaCliqueViewOrganica:
          type: number
          format: double
        beneficiosConsiderados:
          type: array
          items:
            type: string
        promessaResultadoGarantido:
          type: boolean
          const: false
        gratuitoLimitado:
          type: boolean
          const: false
        mensagemSegura:
          type: string
          description: Texto sem promessa de resultado garantido.
    AdminDesempenhoAnuncio:
      type: object
      additionalProperties: false
      properties:
        anuncioId:
          type: string
          format: uuid
        slug:
          type:
            - string
            - "null"
        titulo:
          type:
            - string
            - "null"
        status:
          type:
            - string
            - "null"
        premiumAtivo:
          type: boolean
        visualizacoesTotal:
          type: integer
        cliquesWhatsappTotal:
          type: integer
        taxaCliqueView:
          type: number
          format: double
        periodoInicio:
          type:
            - string
            - "null"
          format: date
        periodoFim:
          type:
            - string
            - "null"
          format: date
        diario:
          type: array
          items:
            $ref: "#/components/schemas/AdminDesempenhoDiario"
        origens:
          type: array
          items:
            $ref: "#/components/schemas/AdminDesempenhoOrigem"
        comparativoPremium:
          $ref: "#/components/schemas/AdminDesempenhoComparativoPremium"
        somenteLeitura:
          type: boolean
          const: true
        dadosSensiveisOcultos:
          type: boolean
          const: true
        trackingExternoExecutado:
          type: boolean
          const: false
    AdminDesempenhoAnunciante:
      type: object
      additionalProperties: false
      properties:
        usuarioId:
          type: string
          format: uuid
        anunciosTotal:
          type: integer
        anunciosComMetricas:
          type: integer
        visualizacoesTotal:
          type: integer
        cliquesWhatsappTotal:
          type: integer
        taxaCliqueView:
          type: number
          format: double
        anuncios:
          type: array
          items:
            $ref: "#/components/schemas/AdminDesempenhoAnuncio"
        somenteLeitura:
          type: boolean
          const: true
        dadosSensiveisOcultos:
          type: boolean
          const: true
        endpointAnuncianteRealDisponivel:
          type: boolean
          const: false
    AdminDesempenhoResumo:
      type: object
      additionalProperties: false
      properties:
        anunciosComMetricas:
          type: integer
        visualizacoesTotal:
          type: integer
        cliquesWhatsappTotal:
          type: integer
        taxaCliqueView:
          type: number
          format: double
        visualizacoesComPremium:
          type: integer
        visualizacoesOrganicas:
          type: integer
        cliquesComPremium:
          type: integer
        cliquesOrganicos:
          type: integer
        somenteLeitura:
          type: boolean
          const: true
        dadosSensiveisOcultos:
          type: boolean
          const: true
        trackingExternoExecutado:
          type: boolean
          const: false
        promessaResultadoGarantido:
          type: boolean
          const: false
        gratuitoLimitado:
          type: boolean
          const: false
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 500
        error:
          type: string
          example: Internal Server Error
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        message:
          type: string
          example: Erro interno inesperado.
        path:
          type: string
          example: /api/health
        requestId:
          type: string
          example: 11111111-1111-4111-8111-111111111111
    ApiErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - UNPROCESSABLE_ENTITY
        - BAD_GATEWAY
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
...